+++ editor/editor-core/src/__tests__/visual-regression/media/layouts.ts
@@ -45,7 +45,7 @@ describe('Snapshot Test: Media', () => {
       ]);
     });
 
-    it('can switch layouts on media', async () => {
+    it.skip('can switch layouts on media', async () => {
       // now we can insert media as necessary
       await insertMedia(page);
       await waitForMediaToBeLoaded(page);
@@ -62,7 +62,7 @@ describe('Snapshot Test: Media', () => {
       }
     });
 
-    it('can switch layouts on individual media', async () => {
+    it.skip('can switch layouts on individual media', async () => {
       // We need a bigger height to capture multiple large images in a row.
       await page.setViewport({ width: 1280, height: 1024 * 2 });