+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-mobile-bridge
 
+## 26.3.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 26.3.5
 
 ### Patch Changes